requirements.txt
# Core agent framework
langgraph==1.1.6
langchain==1.2.15
langchain-core==1.2.28
langchain-groq==1.1.2
langchain-community==0.4.1
langsmith==0.7.30

# LLM provider
groq==0.37.1

# Web search
ddgs==9.13.0

# API
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
slowapi==0.1.9

# Utilities
python-dotenv==1.2.2
requests==2.33.1
httpx==0.28.1
python-multipart==0.0.20

# RAG (optional — PDF upload + Qdrant + local embeddings; torch is pulled in by sentence-transformers)
qdrant-client==1.17.1
sentence-transformers==5.4.1
pypdf==6.10.2

# Testing (Fix #7)
pytest==8.3.5
